adding forestEnemy and forestSpawner class, making all forest enemies spawn from forestSpawner
