+++ Assets/Prefabs/Spawner.prefab
@@ -40,9 +40,12 @@ MonoBehaviour:
   m_GameObject: {fileID: 5822627054572535307}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 27db1f8f39971e54cb693bd19451a5a1, type: 3}
+  m_Script: {fileID: 11500000, guid: 7b29ffd89d7ad0a40a14b31d33a9ea0b, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  spawnCap: 20
+  spawnCap: 5
   Range: 50
-  f1: {fileID: 8428823582346919707, guid: f6b38aedd494f8741a0341bbc16ef870, type: 3}
+  enemy: {fileID: 8428823582346919707, guid: f6b38aedd494f8741a0341bbc16ef870, type: 3}
+  forest:
+  - {fileID: 8428823582346919707, guid: f6b38aedd494f8741a0341bbc16ef870, type: 3}
+  - {fileID: 4083383024262006830, guid: dc669a68a63ed6a44b5b98fc95e2bdfd, type: 3}
