+++ Assets/Prefabs/sunny2.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: dc669a68a63ed6a44b5b98fc95e2bdfd
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
