+++ Assets/Scripts/Enemy/ForestEnemy.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2ee060ac6df8b934b82b7fad03da24d5
+guid: 03e95e2848648394eb40966d666d6ac5
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
