+++ Assets/Scripts/Enemy/Sunny.cs
@@ -1,0 +1,20 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Sunny : ForestEnemy
+{
+    // Start is called before the first frame update
+    public Sunny()
+    {
+        Health = 50;
+        MovementSpeed1 = 6;
+        EnemyName1 = "sunny";
+    }
+
+    // Update is called once per frame
+    void Start()
+    {
+        
+    }
+}
