+++ Assets/Scripts/Enemy/Sunny.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0f641bcd1ae36a04580c3c65e77ccd7e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
