+++ ProjectSettings/TagManager.asset
@@ -6,6 +6,8 @@ TagManager:
   tags:
   - Ground
   - Water
+  - Attack
+  - Effect
   layers:
   - Default
   - TransparentFX
